Moving mocks directory
