+++ test/mocks/generate.go
@@ -6,7 +6,6 @@ package mocks
 //go:generate go-mockgen -f -i Watcher -i ReceiptResponse -o interfaces_transaction.mockgen.go ../../layer1/transaction
 //go:generate go-mockgen -f -i DepositHandler -i AdminHandler -i AdminClient -o interfaces_monitor.mockgen.go ../../layer1/monitor/interfaces
 //go:generate go-mockgen -f -i Client -i AllSmartContracts -i EthereumContracts -o interfaces_ethereum.mockgen.go ../../layer1
-//go:generate go-mockgen -f -i StorageGetter -o interfaces_dynamics.mockgen.go ../../dynamics
 
 // Mocks created from bindings:
